chore(renderer): Created new module BpmnRenderUtil
