+++ lib/draw/BpmnRenderer.js
@@ -3,21 +3,32 @@
 var inherits = require('inherits'),
     isObject = require('lodash/lang/isObject'),
     assign = require('lodash/object/assign'),
-    forEach = require('lodash/collection/forEach'),
-    every = require('lodash/collection/every'),
-    some = require('lodash/collection/some');
+    forEach = require('lodash/collection/forEach');
 
 var BaseRenderer = require('diagram-js/lib/draw/BaseRenderer'),
     TextUtil = require('diagram-js/lib/util/Text'),
     DiUtil = require('../util/DiUtil');
 
-var getBusinessObject = require('../util/ModelUtil').getBusinessObject,
-    is = require('../util/ModelUtil').is;
+var is = require('../util/ModelUtil').is;
 
-var RenderUtil = require('diagram-js/lib/util/RenderUtil');
+var createLine = require('diagram-js/lib/util/RenderUtil').createLine;
 
-var componentsToPath = RenderUtil.componentsToPath,
-    createLine = RenderUtil.createLine;
+var BpmnRenderUtil = require('../util/BpmnRenderUtil');
+
+// helper functions
+var isTypedEvent = BpmnRenderUtil.isTypedEvent,
+    isThrowEvent = BpmnRenderUtil.isThrowEvent,
+    isCollection = BpmnRenderUtil.isCollection,
+    getDi = BpmnRenderUtil.getDi,
+    getSemantic = BpmnRenderUtil.getSemantic;
+
+// cropping path customizations
+var getCirclePath = BpmnRenderUtil.getCirclePath,
+    getRoundRectPath = BpmnRenderUtil.getRoundRectPath,
+    getDiamondPath = BpmnRenderUtil.getDiamondPath,
+    getRectPath = BpmnRenderUtil.getRectPath,
+    getFillColor = BpmnRenderUtil.getFillColor,
+    getStrokeColor = BpmnRenderUtil.getStrokeColor;
 
 var domQuery = require('min-dom/lib/query');
 
@@ -1806,139 +1817,3 @@ BpmnRenderer.prototype.getShapePath = function(element) {
 
   return getRectPath(element);
 };
-
-
-///////// helper functions /////////////////////////////
-
-/**
- * Checks if eventDefinition of the given element matches with semantic type.
- *
- * @return {boolean} true if element is of the given semantic type
- */
-function isTypedEvent(event, eventDefinitionType, filter) {
-
-  function matches(definition, filter) {
-    return every(filter, function(val, key) {
-
-      // we want a == conversion here, to be able to catch
-      // undefined == false and friends
-      /* jshint -W116 */
-      return definition[key] == val;
-    });
-  }
-
-  return some(event.eventDefinitions, function(definition) {
-    return definition.$type === eventDefinitionType && matches(event, filter);
-  });
-}
-
-function isThrowEvent(event) {
-  return (event.$type === 'bpmn:IntermediateThrowEvent') || (event.$type === 'bpmn:EndEvent');
-}
-
-function isCollection(element) {
-  var dataObject = element.dataObjectRef;
-
-  return element.isCollection || (dataObject && dataObject.isCollection);
-}
-
-function getDi(element) {
-  return element.businessObject.di;
-}
-
-function getSemantic(element) {
-  return element.businessObject;
-}
-
-
-
-/////// cropping path customizations /////////////////////////
-
-function getCirclePath(shape) {
-
-  var cx = shape.x + shape.width / 2,
-      cy = shape.y + shape.height / 2,
-      radius = shape.width / 2;
-
-  var circlePath = [
-    ['M', cx, cy],
-    ['m', 0, -radius],
-    ['a', radius, radius, 0, 1, 1, 0, 2 * radius],
-    ['a', radius, radius, 0, 1, 1, 0, -2 * radius],
-    ['z']
-  ];
-
-  return componentsToPath(circlePath);
-}
-
-function getRoundRectPath(shape, borderRadius) {
-
-  var x = shape.x,
-      y = shape.y,
-      width = shape.width,
-      height = shape.height;
-
-  var roundRectPath = [
-    ['M', x + borderRadius, y],
-    ['l', width - borderRadius * 2, 0],
-    ['a', borderRadius, borderRadius, 0, 0, 1, borderRadius, borderRadius],
-    ['l', 0, height - borderRadius * 2],
-    ['a', borderRadius, borderRadius, 0, 0, 1, -borderRadius, borderRadius],
-    ['l', borderRadius * 2 - width, 0],
-    ['a', borderRadius, borderRadius, 0, 0, 1, -borderRadius, -borderRadius],
-    ['l', 0, borderRadius * 2 - height],
-    ['a', borderRadius, borderRadius, 0, 0, 1, borderRadius, -borderRadius],
-    ['z']
-  ];
-
-  return componentsToPath(roundRectPath);
-}
-
-function getDiamondPath(shape) {
-
-  var width = shape.width,
-      height = shape.height,
-      x = shape.x,
-      y = shape.y,
-      halfWidth = width / 2,
-      halfHeight = height / 2;
-
-  var diamondPath = [
-    ['M', x + halfWidth, y],
-    ['l', halfWidth, halfHeight],
-    ['l', -halfWidth, halfHeight],
-    ['l', -halfWidth, -halfHeight],
-    ['z']
-  ];
-
-  return componentsToPath(diamondPath);
-}
-
-function getRectPath(shape) {
-  var x = shape.x,
-      y = shape.y,
-      width = shape.width,
-      height = shape.height;
-
-  var rectPath = [
-    ['M', x, y],
-    ['l', width, 0],
-    ['l', 0, height],
-    ['l', -width, 0],
-    ['z']
-  ];
-
-  return componentsToPath(rectPath);
-}
-
-function getFillColor(element, defaultColor) {
-  var bo = getBusinessObject(element);
-
-  return bo.di.get('fill') || defaultColor || 'white';
-}
-
-function getStrokeColor(element, defaultColor) {
-  var bo = getBusinessObject(element);
-
-  return bo.di.get('stroke') || defaultColor || 'black';
-}
